// front-end/pwa/src/app/metadata/export-specifications/models/raw-export-parameters.model.ts
export interface RawExportParametersModel {
  
  convertDatetimeToDisplayTimeZone?: boolean;
 
  splitObservationDatetime?: boolean;

  unstackData?: boolean;
 
  includeLevel?: boolean;

  includeInterval?: boolean;
 
  includeFlag?: boolean;
 
  includeQCStatus?: boolean;

  includeQCTestLog?: boolean;

  includeComments?: boolean;
 
  includeEntryDatetime?: boolean;

  includeEntryUserEmail?: boolean;

  includeStationName?: boolean;

 
  includeStationLocation?: boolean;
 
  includeStationElevation?: boolean;
 
  includeElementAbbreviation?: boolean;

 
  includeElementName?: boolean;
 
  includeElementUnits?: boolean;

 
  includeSourceName?: boolean;
}